Add some types to models

+++ src/models/qt/editor_tree_item.py
@@ -1,31 +1,40 @@
-from PySide2 import QtWidgets
+from __future__ import annotations
+from typing import Optional
+from PySide2 import QtWidgets, QtGui
+from models.data.editor_item import EditorItem
 
-class EditorTreeItem(object):
-    def __init__(self, label, editor_item=None, is_dir=False, parent=None):
-        self.parent = parent
+class EditorTreeItem:
+    label: str
+    is_dir: bool
+    editor_item: Optional[EditorItem]
+    childItems: list[EditorTreeItem]
+    parent: Optional[EditorTreeItem]
+
+    def __init__(self, label: str, editor_item: Optional[EditorItem] = None, is_dir: bool = False):
         self.label = label
         self.is_dir = is_dir
         self.editor_item = editor_item
         self.childItems = []
+        self.parent = None
 
     @classmethod
-    def from_editor_item(cls, editor_item):
+    def from_editor_item(cls, editor_item: EditorItem) -> EditorTreeItem:
         is_dir = (editor_item.item_type == 'dir')
         return cls(editor_item.name, editor_item, is_dir)
 
-    def data(self):
+    def data(self) -> str:
         return self.label
 
-    def child(self, row):
+    def child(self, row: int) -> Optional[EditorTreeItem]:
         try:
             return self.childItems[row]
         except IndexError:
             return None
 
-    def childCount(self):
+    def childCount(self) -> int:
         return len(self.childItems)
 
-    def childNumber(self):
+    def childNumber(self) -> int:
         if self.parent is not None:
             return self.parent.childItems.index(self)
 
@@ -34,16 +43,18 @@ class EditorTreeItem(object):
     def childIndicatorPolicy(self):
         return QtWidgets.QTreeWidgetItem.ShowIndicator
 
-    def columnCount(self):
+    def columnCount(self) -> int:
         return 1
 
-    def sortChildren(self):
-        self.childItems = sorted(
-            self.childItems, key=lambda x: (not x.is_dir, x.label))
+    def sortChildren(self) -> None:
+        self.childItems = sorted(self.childItems, key=lambda x: (not x.is_dir, x.label))
 
-    def insertChild(self, child_item):
+    def insertChild(self, child_item: EditorTreeItem) -> None:
         self.childItems.append(child_item)
         child_item.parent = self
+
+        if child_item.editor_item is None:
+            return
 
         if self.editor_item is not None:
             child_item.editor_item.parent_id = self.editor_item.id
@@ -54,22 +65,24 @@ class EditorTreeItem(object):
 
         self.sortChildren()
 
-    def insertChildren(self, child_items):
+    def insertChildren(self, child_items: list[EditorTreeItem]) -> None:
         for item in child_items:
             self.insertChild(item)
 
-    def removeChildren(self, position, count, delete=False):
+    def removeChildren(self, position: int, count: int, delete: bool = False) -> bool:
         if position < 0 or position + count > len(self.childItems):
             return False
 
         for row in range(count):
             removed_item = self.childItems.pop(position)
-            if delete:
-                removed_item.editor_item.delete_everything()
+            if delete and self.editor_item is not None:
+                removed_item.editor_item.delete_everything()  # type: ignore
 
         return True
 
-    def setLabel(self, label):
+    def setLabel(self, label: str) -> bool:
+        if self.editor_item is None:
+            return True
         # TODO: Remove *'s becuase they will mess with the modified indicator
         self.label = label
         self.editor_item.name = label
@@ -77,8 +90,8 @@ class EditorTreeItem(object):
 
         return True
 
-    def icon(self):
-        if self.is_dir:
+    def icon(self) -> Optional[QtGui.QIcon]:
+        if self.is_dir or self.editor_item is None:
             return None
         else:
             return self.editor_item.icon()
